+++ mshell-crates/mshell-settings/src/lib.rs
@@ -21,7 +21,7 @@ mod display_settings;
 mod dock_settings;
 mod fonts_settings;
 mod general_settings;
-mod gestures_settings;
+mod input_settings;
 mod weather_settings;
 mod idle_settings;
 mod launcher_settings;